+++ sample-web/src/wasmJsMain/kotlin/com/aditlal/sampleweb/Main.kt
@@ -1,0 +1,11 @@
+package com.aditlal.sampleweb
+
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.CanvasBasedWindow
+
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    CanvasBasedWindow(canvasElementId = "ComposeTarget") {
+        DashboardApp()
+    }
+}